backend/requirements.txt
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Video Processing
yt-dlp>=2024.1.0
scenedetect>=0.6.3
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
pillow>=10.1.0

# Google APIs
google-generativeai>=0.3.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# HTTP & API Clients
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
